implements folium_legend_system and value_mapping_system
