+++ mescal/typevars.py
@@ -11,6 +11,10 @@ if TYPE_CHECKING:
     from mescal.kpis.aggs import OperationOfTwoValues
     from mescal.kpis.kpi_collection import KPICollection
     from mescal.visualizations.folium_viz_system.base_viz_system import ResolvedFeature, FeatureResolver
+    from mescal.visualizations.value_mapping_system.base import BaseMapping
+    from mescal.visualizations.value_mapping_system.discrete import DiscreteInputMapping
+    from mescal.visualizations.value_mapping_system.continuous import SegmentedContinuousInputToContinuousOutputMapping
+    from mescal.visualizations.folium_legend_system.base import BaseLegend
 
 FlagType = TypeVar('FlagType', bound='FlagTypeProtocol')
 DatasetType = TypeVar('DatasetType', bound='Dataset')
@@ -20,3 +24,7 @@ KPIType = TypeVar('KPIType', bound='KPI')
 ValueOperationType = TypeVar('ValueOperationType', bound='OperationOfTwoValues')
 FeatureResolverType = TypeVar('FeatureResolverType', bound='FeatureResolver')
 ResolvedFeatureType = TypeVar('ResolvedFeatureType', bound='ResolvedFeature')
+ValueMappingType = TypeVar('ValueMappingType', bound='BaseMapping')
+DiscreteMappingType = TypeVar('DiscreteMappingType', bound='DiscreteInputMapping')
+ContinuousMappingType = TypeVar('ContinuousMappingType', bound='SegmentedContinuousInputToContinuousOutputMapping')
+FoliumLegendType = TypeVar('FoliumLegendType', bound='BaseLegend')
