+++ mescal/visualizations/value_mapping_system/base.py
@@ -1,0 +1,8 @@
+from abc import ABC, abstractmethod
+from typing import Any
+
+
+class BaseMapping(ABC):
+    @abstractmethod
+    def __call__(self, value) -> Any:
+        pass
